+++ src/App.vue
@@ -47,6 +47,13 @@
       </div>
     </div>
   </div>
+
+  <Toast />
+  <ConfirmDialog>
+    <template #message="{ message: confirmOptions }">
+      <span class="whitespace-pre">{{ confirmOptions.message }}</span>
+    </template>
+  </ConfirmDialog>
 </template>
 
 <style>
@@ -66,6 +73,9 @@
   import { provide, ref } from 'vue';
   import type { useScroll } from '@vueuse/core';
   import { vScroll } from '@vueuse/components';
+
+  import Toast from 'primevue/toast';
+  import ConfirmDialog from 'primevue/confirmdialog';
 
   import NavigationBar from './components/NavigationBar.vue';
   import TopNavigation from './components/TopNavigation.vue';