+++ src/plugins/Plugin.ts
@@ -1,5 +1,7 @@
+import type { PluginInitOptions } from './PluginManager';
+
 export abstract class Plugin {
   public abstract getPluginId(): string;
 
-  public abstract init(): Promise<void>;
+  public abstract init({ registerSettings }: PluginInitOptions): Promise<void>;
 }