+++ src/plugins/PluginManager.ts
@@ -1,7 +1,10 @@
+import type { SettingsConfig } from '../types/SettingsTypes';
 import type { Plugin } from './Plugin';
 
 export class PluginManager {
-  private plugins: Array<Plugin> = [];
+  private readonly plugins: Array<Plugin> = [];
+
+  private readonly settings: Array<SettingsConfig> = [];
 
   constructor(...plugins: Array<Plugin>) {
     this.plugins.push(...plugins);
@@ -9,7 +12,11 @@ export class PluginManager {
 
   public async init(): Promise<void> {
     for (const plugin of this.plugins) {
-      await plugin.init();
+      await plugin.init({
+        registerSettings: (settings) => {
+          this.settings.push(...settings);
+        }
+      });
     }
   }
 
@@ -18,4 +25,12 @@ export class PluginManager {
   ): T | null {
     return this.plugins.find((p) => p.constructor === pluginClass) ?? null;
   }
+
+  public getSettings(): Array<SettingsConfig> {
+    return this.settings;
+  }
 }
+
+export type PluginInitOptions = {
+  registerSettings: (settings: Array<SettingsConfig>) => void;
+};