+++ src/stores/Store.ts
@@ -26,6 +26,8 @@ export class Store<
   private notifyRemoved;
   private notifyUpserted;
 
+  private migrationHelper;
+
   protected readonly initializePromise;
 
   protected constructor({
@@ -46,6 +48,8 @@ export class Store<
     this.notifyRemoved = notifyEntityRemoved;
     this.notifyUpserted = notifyEntityUpserted;
 
+    this.migrationHelper = new MigrationHelper(this.tableName);
+
     this.initializePromise = migrationConfig
       ? this.migrate(migrationConfig)
       : Promise.resolve();
@@ -59,6 +63,11 @@ export class Store<
   public async getAll(): Promise<Array<TEntity>> {
     await this.initializePromise;
     return this.store.many({});
+  }
+
+  public async importData(entities: Array<TEntity>): Promise<void> {
+    await this.clear();
+    await this.store.upsertMany(entities);
   }
 
   protected _watchForComputedQuery(
@@ -221,12 +230,14 @@ export class Store<
     }
   }
 
+  public async getStoreVersion(): Promise<number> {
+    return this.migrationHelper.getLastDbVersion();
+  }
+
   private async migrate(
     migrationConfig: MigrationConfig<TEntity, TRenamedProperties>
   ): Promise<void> {
-    const migrationHelper = new MigrationHelper(this.tableName);
-
-    const lastDbVersion = await migrationHelper.getLastDbVersion();
+    const lastDbVersion = await this.migrationHelper.getLastDbVersion();
     const newDbVersion = migrationConfig.version;
 
     if (newDbVersion < lastDbVersion) {
@@ -238,7 +249,7 @@ export class Store<
     const migratedEntities = migrationConfig.migrationFunction(entities);
     await this.store.upsertMany(migratedEntities);
 
-    await migrationHelper.setLastDbVersion(migrationConfig.version);
+    await this.migrationHelper.setLastDbVersion(migrationConfig.version);
   }
 }